pyproject.toml
[tool.poetry]
name = "cars-project"
version = "0.1.0"
description = ""
authors = ["Grigoriy1209 <grigoriyv1209@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
djangorestframework = "^3.15.2"
mysqlclient = "^2.2.5"
djangorestframework-simplejwt = "^5.3.1"



[tool.poetry.group.dev.dependencies]
django-dotenv = "^1.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
